feat: refactor theme handling and add user provider for authentication
